+++ src/backend/resolvers/mutations/indication/update.js
@@ -1,67 +1,102 @@
-const _ = 'lodash'
+// TODO: Deprecate old Phoenix Treatment Plans Panel; this resolver is bad
+
 const { ObjectId } = require('mongodb')
+const getIndTherapeuticAreaPipeline = require('./getIndTherapeuticAreaPipeline')
 
 const updateSourceIndication = async (
   parent,
-  { input: { _id: indicationId, ...body } },
+  { input: { _id, ...body } },
   { pulseCoreDb, pulseDevDb, mongoClient },
-  info,
+  info
 ) => {
-  const _id = ObjectId(indicationId)
+  _id = ObjectId(_id)
 
-  if (body.regimens) { // TODO: shouldn't have to guard against this; every indication should have an regimens array
-    const editedRegimens = body.regimens.map(({ _id: regimenId, name, products }) => {
-      const newRegimenId = ObjectId(regimenId)
+  // ! if updating is coming from the indications panel, regimens won't be present at all
+  // ! and that slice of the indication doc must remain untouched, so don't add that key to body;
+  // ! otherwise, updating is coming from Phoenix Treatment Plans panel, and we need to
+  // ! perform the following logic
+  if (body.regimens) {
+    const editedRegimens = body.regimens.map(
+      ({ _id: regimenId, name, products }) => {
+        const newRegimenId = ObjectId(regimenId)
 
-      const editedProducts = products.map(({ _id: productId, ...product }) => {
-        const newProductId = ObjectId(productId)
-        return { _id: newProductId, ...product }
-      })
+        const editedProducts = products.map(
+          ({ _id: productId, ...product }) => {
+            const newProductId = ObjectId(productId)
+            return { _id: newProductId, ...product }
+          }
+        )
 
-      return {
-        _id: newRegimenId,
-        name,
-        products: editedProducts,
+        return {
+          _id: newRegimenId,
+          name,
+          products: editedProducts,
+        }
       }
-    })
+    )
 
     body.regimens = editedRegimens
   }
 
+  // ! if 'therapeuticAreaId' key exists in incoming body, that means update op
+  // ! is coming from indications panel, and we should throw an error if the user
+  // ! hasn't selected a corresponding therapeutic area;
+  // ! otherwise, the update is coming from the phoenix treatment plans (ind+reg)
+  // ! panel and that panel has no bearing on the 'therapeuticAreaId' slice, so
+  // ! we leave 'therapeuticAreaId' out of what gets $set in the update op
+  if ('therapeuticAreaId' in body) {
+    if (!body.therapeuticAreaId) {
+      throw new Error(`Therapeutic area must be selected for the indication`)
+    } else {
+      body.therapeuticAreaId = ObjectId(body.therapeuticAreaId)
+    }
+  }
+
   const session = mongoClient.startSession()
 
-  let result
+  let updatedIndication
   await session.withTransaction(async () => {
-    const { value: updatedIndication } = await pulseCoreDb
+    // Step 1: Update the indication
+    updatedIndication = await pulseCoreDb
       .collection('indications')
       .findOneAndUpdate(
         { _id },
         { $set: body },
-        { returnOriginal: false, session },
+        { returnOriginal: false, session }
       )
+      .then(({ value }) => value)
+
+    // Step 2: Update the indication's name in pulse-dev.users.nodes.resources
+    await pulseDevDb.collection('users.nodes.resources').updateMany(
+      { 'resources.treatmentPlans._id': _id },
+      {
+        $set: {
+          'resources.$[resource].treatmentPlans.$[indication].name':
+            updatedIndication.name,
+        },
+      },
+      {
+        arrayFilters: [
+          { 'resource.treatmentPlans': { $exists: true } },
+          { 'indication._id': _id },
+        ],
+        session,
+      }
+    )
+
+    // Step 3: Materialize the updated indication/therapeuticArea combo doc and update
+    // corresponding doc in pulse-dev.indicationsTherapeuticAreas collection
+    const docToUpdate = await pulseCoreDb
+      .collection('indications')
+      .aggregate(getIndTherapeuticAreaPipeline(_id), { session })
+      .next()
 
     await pulseDevDb
-      .collection('users.nodes.resources')
-      .updateMany(
-        { 'resources.treatmentPlans._id': _id },
-        {
-          $set: {
-            'resources.$[resource].treatmentPlans.$[indication].name': updatedIndication.name
-          }
-        },
-        {
-          arrayFilters: [
-            { 'resource.treatmentPlans': { $exists: true } },
-            { 'indication._id': _id }
-          ],
-          session,
-        }
-      )
-  
-    result = updatedIndication
+      .collection('indicationsTherapeuticAreas')
+      .updateOne({ _id }, { $set: docToUpdate }, { session, upsert: true })
   })
 
-  return result
+  return updatedIndication
 }
 
 module.exports = updateSourceIndication