+++ src/backend/resolvers/mutations/therapeuticArea/index.js
@@ -1,0 +1,9 @@
+const createTherapeuticArea = require('./create')
+const updateTherapeuticArea = require('./update')
+const deleteTherapeuticArea = require('./delete')
+
+module.exports = {
+  createTherapeuticArea,
+  updateTherapeuticArea,
+  deleteTherapeuticArea,
+}