+++ src/backend/resolvers/queries/therapeuticAreas.js
@@ -1,0 +1,4 @@
+const therapeuticAreas = (parent, args, { pulseCoreDb }, info) =>
+  pulseCoreDb.collection('therapeuticAreas').find().toArray()
+
+module.exports = therapeuticAreas