+++ src/backend/typeDefs/mutations/therapeuticArea/create.js
@@ -1,0 +1,14 @@
+const { gql } = require('apollo-server-express')
+
+const createTherapeuticAreaTypeDefs = gql`
+  input CreateTherapeuticAreaInput {
+    name: String!
+  }
+
+  type CreateTherapeuticAreaPayload {
+    _id: ID!
+    name: String
+  }
+`
+
+module.exports = createTherapeuticAreaTypeDefs