+++ src/backend/typeDefs/mutations/therapeuticArea/index.js
@@ -1,0 +1,9 @@
+const createTherapeuticAreaTypeDefs = require('./create')
+const updateTherapeuticAreaTypeDefs = require('./update')
+const deleteTherapeuticAreaTypeDefs = require('./delete')
+
+module.exports = [
+  createTherapeuticAreaTypeDefs,
+  updateTherapeuticAreaTypeDefs,
+  deleteTherapeuticAreaTypeDefs,
+]