+++ src/frontend/Orion/Sidebar.js
@@ -120,6 +120,10 @@ const PRODUCT_INDICATION_LINKS_CONFIG = [
     link: '/orion/configuration/products-indications/indications',
   },
   {
+    label: 'Therapeutic Areas',
+    link: '/orion/configuration/products-indications/therapeutic-areas',
+  },
+  {
     label: 'Regimens',
     link: '/orion/configuration/products-indications/regimens',
   },