+++ src/frontend/hooks/useBom/boid-query-map.js
@@ -7,6 +7,7 @@ import {
   GET_OBM_SERVICES,
   GET_OBM_SERVICES_CATEGORIES,
   GET_PEOPLE,
+  GET_THERAPEUTIC_AREAS,
 } from './../../api/queries'
 
 export default {
@@ -15,4 +16,5 @@ export default {
   '5ed81e5fb8ebf33703463750': GET_OBM_SERVICES,
   '5ed81ed8bcfdf6381562c17e': GET_OBM_SERVICES_CATEGORIES,
   '5eea22d5adbf920fa4320487': GET_PEOPLE,
+  '5f2ac32600b9d74431f9bc6f': GET_THERAPEUTIC_AREAS,
 }