+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `Phoenix/Backend`: Write integration and unit tests for `upsertUsersSitemaps.js` util
+- `Orion`: Build out therapeuticArea CRUD and integrate with indications master list CRUD; build out all cascade logic including syncing with materialized `pulse-dev.indicationsTherapeuticAreas`
 
 ## [3.2.0] - Started: 7-24-20